refactor: move background-job bootstrap out of the data layer
setupDatabase carried a ~150-line composition root: the whole Ankify subsystem (RacService + Docker, export scheduler, sync use case, polling) was wired inside src/data_layer/index.ts, making the DB layer import use cases and services upward. That inversion produced 5 data-layer-is-leaf violations and all 7 index.ts dependency cycles, and meant every route importing getDatabase transitively loaded dockerode, the Notion client, and EmailService at module load. The wiring moves verbatim to src/app/bootstrapBackgroundJobs.ts, called from server.ts right after setupDatabase.

Also fixes a live double-schedule: ScheduleCleanup ran twice on the prod main instance (server.ts unconditionally + data_layer/index.ts behind its prod guard), and the Singapore disable branch was dead because server.ts scheduled it there anyway. Cleanup now schedules exactly once, behind the singapore guard, with a source-text regression test pinning both properties.

Behavior deltas, intended: cleanup once instead of twice on prod; no cleanup on singapore (the explicitly-intended behavior); ankify bootstrap now runs after failStrandedLegacyJobs instead of before (independent tables, order immaterial); a bootstrap failure now also lands in the fallback error file via serve().catch before the same exit(1). Everything else is a verbatim move. Deliberately left: the one type-only ShowcaseRepository leaf warning (its fix requires relocating PreviewBlockPayload out of controllers - separate change).

Full server suite 6606 passed (only the documented pdfinfo environmental failures); tsc clean; depcruise 319 -> 298 warnings with data-layer-is-leaf 5 -> 1.
